Remove dead code, update and retain tests
